Drop _closeButton; Header defines the static close-button appearance
The close button's look (× icon, tertiary variant, localized aria-label) never
varies, so there's no need to pass it through context. Header provides it
directly on the close slot; only the dynamic title id still comes from Dialog via
HeaderContext. content.tsx now imports the Close icon + locale for this.
